requirements.txt
python-telegram-bot>=20.0
apscheduler>=3.10.0
motor>=3.3.0
pymongo>=4.6.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
httpx>=0.25.0
python-dotenv>=1.0.0
pytz>=2023.3
pydantic>=2.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
